rename static to pure component, small fixes
