+++ components/button/element.js
@@ -1,7 +1,7 @@
 import { LitElement, html } from "lit";
 
+import { Button } from "./pure.js";
 import styles from "./shadow.css?lit";
-import { Button } from "./static.js";
 
 export class MDNButton extends LitElement {
   static styles = styles;
